Generalize security code verification cards
