+++ frontend/src/app/accountCreation/MfaSmsVerify/MfaSmsVerify.stories.tsx
@@ -6,9 +6,12 @@ export default {
   title: "App/Account set up/Step 3b: Verify SMS security code",
   component: MfaSmsVerify,
   argTypes: {},
+  args: { phoneNumber: "(530) 867-5309" },
 } as Meta;
 
-const Template: Story = (args) => <MfaSmsVerify {...args} />;
+type Props = React.ComponentProps<typeof MfaSmsVerify>;
+
+const Template: Story<Props> = (args) => <MfaSmsVerify {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {};
