+++ frontend/src/app/accountCreation/MfaSmsVerify/MfaSmsVerify.tsx
@@ -1,66 +1,16 @@
-import { useState } from "react";
+import { MfaVerify } from "../MfaVerify/MfaVerify";
 
-import { Card } from "../../commonComponents/Card/Card";
-import { CardBackground } from "../../commonComponents/CardBackground/CardBackground";
-import TextInput from "../../commonComponents/TextInput";
-import Button from "../../commonComponents/Button/Button";
-import StepIndicator from "../../commonComponents/StepIndicator";
-import { accountCreationSteps } from "../../../config/constants";
+interface Props {
+  phoneNumber: string;
+}
 
-export const MfaSmsVerify = () => {
-  const [code, setCode] = useState("");
-  const [codeError, setCodeError] = useState("");
-
-  const validateCode = () => {
-    if (code === "") {
-      setCodeError("Enter your security code");
-    } else {
-      setCodeError("");
+export const MfaSmsVerify = (props: Props) => (
+  <MfaVerify
+    hint={
+      <>
+        We’ve sent a text message (SMS) to <b>{props.phoneNumber}</b>. It will
+        expire in 10 minutes.
+      </>
     }
-  };
-
-  return (
-    <CardBackground>
-      <Card logo bodyKicker="Set up your account">
-        <StepIndicator
-          steps={accountCreationSteps}
-          currentStepValue={"2"}
-          noLabels={true}
-        />
-        <p className="margin-bottom-0">Verify your security code.</p>
-        <p className="usa-hint font-ui-2xs margin-bottom-0">
-          We’ve sent a text message (SMS) to <b>(213) 555-2424</b>. It will
-          expire in 10 minutes.{" "}
-        </p>
-        <div className="display-flex">
-          <TextInput
-            className="flex-fill"
-            label={"One-time security code"}
-            name={"security-code"}
-            type={"tel"}
-            required
-            value={code}
-            errorMessage={codeError}
-            validationStatus={codeError ? "error" : undefined}
-            onBlur={validateCode}
-            onChange={(evt) => setCode(evt.currentTarget.value)}
-          />
-          <Button
-            className="margin-top-3 flex-align-self-end margin-left-1"
-            label={"Verify"}
-            type={"submit"}
-            onClick={validateCode}
-          />
-        </div>
-        <Button
-          className="usa-button--outline display-block margin-top-3"
-          label={"Send another code"}
-          type={"submit"}
-        />
-      </Card>
-      <p className="margin-top-5">
-        <a href="#0">Return to previous step</a>
-      </p>
-    </CardBackground>
-  );
-};
+  />
+);
